Refactor: Extract /top command to separate module to fix line limit
Fixes check-file-line-limits CI failure by reducing telegram-bot.mjs
from 1765 lines to 1484 lines (under 1500 line limit).

Changes:
- Created src/telegram-top-command.lib.mjs (325 lines)
- Extracted all /top command logic and stop button handler
- Optimized verbose logging to reduce line count
- Maintained all functionality and experimental warnings

The /top command implementation is now in a dedicated module
with a clean registerTopCommand() API.
